backend/requirements.txt
click==8.1.3
Flask==2.1.2
Flask-API==3.0.post1
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-dotenv==0.20.0
six==1.16.0
SQLAlchemy==1.4.39
Werkzeug==2.1.2
